requirements-local.txt
# PACKAGES APP
jina==2.6.*
torch==1.10.*
transformers==4.16.*
psycopg2-binary==2.9.*
SQLAlchemy==1.4.*
python-dotenv==0.19.*
wordcloud==1.8.1
nltk==3.6.*
passlib[bcrypt]==1.7.*
python-jose[cryptography]==3.3.0
emails==0.6
email-validator==1.1.*

# PACKAGES ENV DEV
autoflake
black
flake8
isort
mypy

pipdeptree
